section 7 complete
Split clonezilla/block_cache.rs (1461 lines) into a cache module and a
metadata_scan module. block_cache.rs (415 lines) keeps PartcloneBlockCache,
PartcloneBlockReader, CacheState, and persistence. metadata_scan.rs (1063
lines) holds scan_metadata plus the per-filesystem identify_*/discover_*
helpers (FAT/NTFS/exFAT/ext/btrfs).

Marked the audit's other §7 items resolved on review: scan_metadata and
backup/metadata.rs parse different inputs (partclone block stream vs JSON
schema) and don't actually overlap; Clonezilla GPT already routes through
partition::gpt::Gpt::parse at restore/mod.rs:1428.
