+++ src/clonezilla/mod.rs
@@ -1,3 +1,4 @@
 pub mod block_cache;
 pub mod metadata;
+pub mod metadata_scan;
 pub mod partclone;
